+++ friendsbets-core-springboot/src/main/java/friendsbets/core/sb/repositories/BetRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import friendsbets.core.sb.models.Bet;
 
-public interface BetRepository extends JpaRepository<Bet, Integer>{
+public interface BetRepository extends JpaRepository<Bet, Long>{
 
 }
