+++ friendsbets-core-springboot/src/main/java/friendsbets/core/sb/repositories/GroupRepository.java
@@ -7,14 +7,14 @@ import org.springframework.data.jpa.repository.Query;
 
 import friendsbets.core.sb.models.Group;
 
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 
 	@Query(value =
 			"SELECT * FROM GroupFbs "
 			+ "INNER JOIN GroupFbs_userList ON GroupFbs.id = GroupFbs_userList.grpList_id "
 			+ "WHERE userList_id = ?1"
 			, nativeQuery = true)
-	Set<Group> findAllForOneUser(int id);
+	Set<Group> findAllForOneUser(long id);
